requirements.txt
uwsgi
redis
pyotp
pycrypt
requests==2.19.1
PyMySQL
mysqlclient==1.3.13
pytest-flask
celery==3.1.25
ipfsapi==0.4.3
Flask==0.10.1
Flask-Script
Flask-SqlAlchemy
Flask-Redis==0.3.0
Flask-Script
Flask-Cache
Flask-Restful==0.3.6
Flask-Babel==0.11.2
Flask-Celery-Helper==1.1.0
base58==0.2.5
boto3==1.7.62
pycoin==0.80
validate_email
nexmo
twilio
cashaddress
